+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2025-12-23 16:09:52
+
+### Changed Files
+- `src/services/recruiting/authUserService.ts`
+
+### Statistics
+```
+ 1 file changed, 3 insertions(+), 17 deletions(-)
+```
+
 ## [Checkpoint] - 2025-12-23 16:00:13
 
 ### Changed Files